+++ SidebarFavoritesManager/Models/Config.swift
@@ -9,6 +9,27 @@ struct Config: Codable {
     struct Settings: Codable {
         var launchAtLogin: Bool = false
         var showInMenuBar: Bool = true
+        var signingIdentity: SigningIdentity = .automatic
+
+        // Custom coding keys to handle older config files without signingIdentity
+        enum CodingKeys: String, CodingKey {
+            case launchAtLogin
+            case showInMenuBar
+            case signingIdentity
+        }
+
+        init(launchAtLogin: Bool = false, showInMenuBar: Bool = true, signingIdentity: SigningIdentity = .automatic) {
+            self.launchAtLogin = launchAtLogin
+            self.showInMenuBar = showInMenuBar
+            self.signingIdentity = signingIdentity
+        }
+
+        init(from decoder: Decoder) throws {
+            let container = try decoder.container(keyedBy: CodingKeys.self)
+            launchAtLogin = try container.decodeIfPresent(Bool.self, forKey: .launchAtLogin) ?? false
+            showInMenuBar = try container.decodeIfPresent(Bool.self, forKey: .showInMenuBar) ?? true
+            signingIdentity = try container.decodeIfPresent(SigningIdentity.self, forKey: .signingIdentity) ?? .automatic
+        }
     }
 
     init(favorites: [Favorite] = [], settings: Settings = Settings()) {